+++ handler/handler.go
@@ -1,10 +1,18 @@
 package handler
 
-import "github.com/jinzhu/gorm"
+import (
+	"os/user"
+
+	"github.com/jinzhu/gorm"
+)
 
 // Handler retrun a DB object
 type Handler struct {
 	DB *gorm.DB
+}
+
+type User struct {
+	*user.User
 }
 
 const (
